+++ pennylane_lightning/core/src/simulators/lightning_kokkos/measurements/MeasurementsKokkos.hpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 #pragma once
 #include <chrono>
+#include <cstdint>
 
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Random.hpp>
@@ -28,11 +29,14 @@
 
 /// @cond DEV
 namespace {
+using namespace Pennylane::LightningKokkos::Functors;
 using namespace Pennylane::Measures;
 using namespace Pennylane::Observables;
 using Pennylane::LightningKokkos::StateVectorKokkos;
 using Pennylane::LightningKokkos::Util::getRealOfComplexInnerProduct;
 using Pennylane::LightningKokkos::Util::SparseMV_Kokkos;
+using Pennylane::LightningKokkos::Util::vector2view;
+using Pennylane::LightningKokkos::Util::view2vector;
 using Pennylane::Util::exp2;
 enum class ExpValFunc : uint32_t {
     BEGIN = 1,
@@ -54,8 +58,8 @@ class Measurements final
     using PrecisionT = typename StateVectorT::PrecisionT;
     using ComplexT = typename StateVectorT::ComplexT;
     using BaseType = MeasurementsBase<StateVectorT, Measurements<StateVectorT>>;
-
     using KokkosExecSpace = typename StateVectorT::KokkosExecSpace;
+    using HostExecSpace = typename StateVectorT::HostExecSpace;
     using KokkosVector = typename StateVectorT::KokkosVector;
     using KokkosSizeTVector = typename StateVectorT::KokkosSizeTVector;
     using UnmanagedSizeTHostView =
@@ -64,8 +68,6 @@ class Measurements final
         typename StateVectorT::UnmanagedConstComplexHostView;
     using UnmanagedConstSizeTHostView =
         typename StateVectorT::UnmanagedConstSizeTHostView;
-    using UnmanagedPrecisionHostView =
-        typename StateVectorT::UnmanagedPrecisionHostView;
     using ScratchViewComplex = typename StateVectorT::ScratchViewComplex;
     using TeamPolicy = typename StateVectorT::TeamPolicy;
 
@@ -213,10 +215,7 @@ class Measurements final
         PL_ABORT_IF(matrix_.size() != exp2(2 * wires.size()),
                     "The size of matrix does not match with the given "
                     "number of wires");
-        KokkosVector matrix("matrix_", matrix_.size());
-        Kokkos::deep_copy(matrix, UnmanagedConstComplexHostView(
-                                      matrix_.data(), matrix_.size()));
-        return getExpValMatrix(matrix, wires);
+        return getExpValMatrix(vector2view(matrix_), wires);
     };
 
     /**
@@ -477,26 +476,31 @@ class Measurements final
     /**
      * @brief Probabilities to measure rotated basis states.
      *
+     * @return Kokkos::View with probabilities
+     * in lexicographic order.
+     */
+    auto probs_core() -> Kokkos::View<PrecisionT *> {
+        const std::size_t N = this->_statevector.getLength();
+        auto sv = this->_statevector.getView();
+        Kokkos::View<PrecisionT *> d_probs("d_probs", N);
+        Kokkos::parallel_for(
+            Kokkos::RangePolicy<KokkosExecSpace>(0, N),
+            KOKKOS_LAMBDA(const std::size_t k) {
+                const PrecisionT rsv = sv(k).real();
+                const PrecisionT isv = sv(k).imag();
+                d_probs(k) = rsv * rsv + isv * isv;
+            });
+        return d_probs;
+    }
+
+    /**
+     * @brief Probabilities to measure rotated basis states.
+     *
      * @return Floating point std::vector with probabilities
      * in lexicographic order.
      */
     auto probs() -> std::vector<PrecisionT> {
-        const std::size_t N = this->_statevector.getLength();
-
-        Kokkos::View<ComplexT *> arr_data = this->_statevector.getView();
-        Kokkos::View<PrecisionT *> d_probability("d_probability", N);
-
-        // Compute probability distribution from StateVector
-        Kokkos::parallel_for(
-            Kokkos::RangePolicy<KokkosExecSpace>(0, N),
-            getProbFunctor<PrecisionT>(arr_data, d_probability));
-
-        std::vector<PrecisionT> probabilities(N, 0);
-
-        Kokkos::deep_copy(UnmanagedPrecisionHostView(probabilities.data(),
-                                                     probabilities.size()),
-                          d_probability);
-        return probabilities;
+        return view2vector(probs_core());
     }
 
     /**
@@ -507,119 +511,91 @@ class Measurements final
      * @return Floating point std::vector with probabilities.
      * The basis columns are rearranged according to wires.
      */
-    auto probs(const std::vector<std::size_t> &wires)
+    auto
+    probs(const std::vector<std::size_t> &wires,
+          [[maybe_unused]] const std::vector<std::size_t> &device_wires = {})
         -> std::vector<PrecisionT> {
-        PL_ABORT_IF_NOT(
-            std::is_sorted(wires.cbegin(), wires.cend()),
-            "LightningKokkos does not currently support out-of-order wire "
-            "indices with probability calculations");
-
-        // If all wires are requested, dispatch to `this->probs()`
-        if (wires.size() == this->_statevector.getNumQubits()) {
-            return this->probs();
+        // GPU_SHARED_NWIRES_MAX is an upper bound for the size of the GPU array
+        // used to reduce the probs (max size = 2 ** 7)
+        constexpr std::size_t GPU_SHARED_NWIRES_MAX = 7;
+        // BITSHIFT_FREE_WIRES_MIN is a lower bound for the size of the loop
+        // over which the probs computation is parallelized in
+        // `probs_bitshift_generic` The free wires are the wires which are
+        // summed over.
+        constexpr std::size_t BITSHIFT_FREE_WIRES_MIN = 10;
+        // BITSHIFT_NWIRES_MAX is an upper bound for using
+        // `probs_bitshift_generic`, beyond that size the other implementation
+        // si more efficient
+        constexpr std::size_t BITSHIFT_NWIRES_MAX = 9;
+        // MDRANGE_NWIRES_MAX is an upper bound for using `MDRangePolicy` to
+        // parallelize the probs computation. Beyond that size, parallelizing
+        // over the probs elements is more efficient.
+        constexpr std::size_t MDRANGE_NWIRES_MAX = 20;
+        const std::size_t n_wires = wires.size();
+        if (n_wires == 0) {
+            return {1.0};
         }
-
-        using MDPolicyType_2D =
-            Kokkos::MDRangePolicy<Kokkos::Rank<2, Kokkos::Iterate::Left>>;
-
-        //  Determining probabilities for the sorted wires.
-        const Kokkos::View<ComplexT *> arr_data = this->_statevector.getView();
         const std::size_t num_qubits = this->_statevector.getNumQubits();
-
-        std::vector<std::size_t> sorted_ind_wires(wires);
-        const bool is_sorted_wires =
-            std::is_sorted(sorted_ind_wires.begin(), sorted_ind_wires.end());
-        std::vector<std::size_t> sorted_wires(wires);
-
-        if (!is_sorted_wires) {
-            sorted_ind_wires = Pennylane::Util::sorting_indices(wires);
-            for (std::size_t pos = 0; pos < wires.size(); pos++)
-                sorted_wires[pos] = wires[sorted_ind_wires[pos]];
+        // is_equal_to_all_wires is True if `wires` includes all wires in order
+        // and false otherwise
+        bool is_equal_to_all_wires = n_wires == num_qubits;
+        for (std::size_t k = 0; k < n_wires; k++) {
+            if (!is_equal_to_all_wires) {
+                break;
+            }
+            is_equal_to_all_wires = wires[k] == k;
         }
-
+        if (is_equal_to_all_wires) {
+            return probs();
+        }
+        const bool is_gpu_scratch_limited =
+            n_wires > GPU_SHARED_NWIRES_MAX &&
+            !std::is_same_v<KokkosExecSpace, HostExecSpace>;
+        if (num_qubits - n_wires > BITSHIFT_FREE_WIRES_MIN &&
+            n_wires < BITSHIFT_NWIRES_MAX && !is_gpu_scratch_limited) {
+            return probs_bitshift_generic<KokkosExecSpace>(
+                this->_statevector.getView(), num_qubits, wires);
+        }
         std::vector<std::size_t> all_indices =
-            Pennylane::Util::generateBitsPatterns(sorted_wires, num_qubits);
-
+            Pennylane::Util::generateBitsPatterns(wires, num_qubits);
+        Kokkos::View<std::size_t *> d_all_indices = vector2view(all_indices);
         std::vector<std::size_t> all_offsets =
             Pennylane::Util::generateBitsPatterns(
-                Pennylane::Util::getIndicesAfterExclusion(sorted_wires,
-                                                          num_qubits),
+                Pennylane::Util::getIndicesAfterExclusion(wires, num_qubits),
                 num_qubits);
-
+        Kokkos::View<std::size_t *> d_all_offsets = vector2view(all_offsets);
         Kokkos::View<PrecisionT *> d_probabilities("d_probabilities",
                                                    all_indices.size());
+        Kokkos::View<ComplexT *> sv = this->_statevector.getView();
 
-        Kokkos::View<std::size_t *> d_sorted_ind_wires("d_sorted_ind_wires",
-                                                       sorted_ind_wires.size());
-        Kokkos::View<std::size_t *> d_all_indices("d_all_indices",
-                                                  all_indices.size());
-        Kokkos::View<std::size_t *> d_all_offsets("d_all_offsets",
-                                                  all_offsets.size());
-
-        Kokkos::deep_copy(
-            d_all_indices,
-            UnmanagedSizeTHostView(all_indices.data(), all_indices.size()));
-        Kokkos::deep_copy(
-            d_all_offsets,
-            UnmanagedSizeTHostView(all_offsets.data(), all_offsets.size()));
-        Kokkos::deep_copy(d_sorted_ind_wires,
-                          UnmanagedSizeTHostView(sorted_ind_wires.data(),
-                                                 sorted_ind_wires.size()));
-
-        const int num_all_indices =
-            all_indices.size(); // int is required by Kokkos::MDRangePolicy
-        const int num_all_offsets = all_offsets.size();
-
-        MDPolicyType_2D mdpolicy_2d0({{0, 0}},
-                                     {{num_all_indices, num_all_offsets}});
-
-        Kokkos::parallel_for(
-            "Set_Prob", mdpolicy_2d0,
-            KOKKOS_LAMBDA(const std::size_t i, const std::size_t j) {
-                const std::size_t index = d_all_indices(i) + d_all_offsets(j);
-                const PrecisionT REAL = arr_data(index).real();
-                const PrecisionT IMAG = arr_data(index).imag();
-                const PrecisionT value = REAL * REAL + IMAG * IMAG;
-                Kokkos::atomic_add(&d_probabilities(i), value);
-            });
-
-        std::vector<PrecisionT> probabilities(all_indices.size(), 0);
-
-        if (is_sorted_wires) {
-            Kokkos::deep_copy(UnmanagedPrecisionHostView(probabilities.data(),
-                                                         probabilities.size()),
-                              d_probabilities);
-            return probabilities;
+        // Reducing over `d_probabilities` requires too much L0 scratch memory
+        // on GPUs. If n_wires >= 20, this also requires quite a bit of memory
+        // on CPUs, so we fallback to the next implementation
+        if (n_wires < MDRANGE_NWIRES_MAX && !is_gpu_scratch_limited) {
+            using MDPolicyType_2D =
+                Kokkos::MDRangePolicy<Kokkos::Rank<2, Kokkos::Iterate::Left>>;
+            auto md_policy = MDPolicyType_2D(
+                {{0, 0}}, {{static_cast<int64_t>(all_indices.size()),
+                            static_cast<int64_t>(all_offsets.size())}});
+            Kokkos::parallel_reduce(
+                md_policy,
+                getProbsFunctor<PrecisionT, KokkosExecSpace>(
+                    sv, wires, d_all_indices, d_all_offsets),
+                d_probabilities);
         } else {
-            Kokkos::View<PrecisionT *> transposed_tensor("transposed_tensor",
-                                                         all_indices.size());
-
-            Kokkos::View<std::size_t *> d_trans_index("d_trans_index",
-                                                      all_indices.size());
-
-            const int num_trans_tensor = transposed_tensor.size();
-            const int num_sorted_ind_wires = sorted_ind_wires.size();
-
-            MDPolicyType_2D mdpolicy_2d1(
-                {{0, 0}}, {{num_trans_tensor, num_sorted_ind_wires}});
-
             Kokkos::parallel_for(
-                "TransIndex", mdpolicy_2d1,
-                getTransposedIndexFunctor(d_sorted_ind_wires, d_trans_index,
-                                          num_sorted_ind_wires));
-
-            Kokkos::parallel_for(
-                "Transpose",
-                Kokkos::RangePolicy<KokkosExecSpace>(0, num_trans_tensor),
-                getTransposedFunctor<PrecisionT>(
-                    transposed_tensor, d_probabilities, d_trans_index));
-
-            Kokkos::deep_copy(UnmanagedPrecisionHostView(probabilities.data(),
-                                                         probabilities.size()),
-                              transposed_tensor);
-
-            return probabilities;
+                all_indices.size(), KOKKOS_LAMBDA(const std::size_t i) {
+                    for (std::size_t j = 0; j < d_all_offsets.size(); j++) {
+                        const std::size_t index =
+                            d_all_indices(i) + d_all_offsets(j);
+                        const PrecisionT rsv = sv(index).real();
+                        const PrecisionT isv = sv(index).imag();
+                        d_probabilities(i) += rsv * rsv + isv * isv;
+                    }
+                });
         }
+
+        return view2vector(d_probabilities);
     }
 
     /**
@@ -682,17 +658,11 @@ class Measurements final
     auto generate_samples(std::size_t num_samples) -> std::vector<std::size_t> {
         const std::size_t num_qubits = this->_statevector.getNumQubits();
         const std::size_t N = this->_statevector.getLength();
-
-        Kokkos::View<ComplexT *> arr_data = this->_statevector.getView();
-        Kokkos::View<PrecisionT *> probability("probability", N);
         Kokkos::View<std::size_t *> samples("num_samples",
                                             num_samples * num_qubits);
 
-        // Compute probability distribution from StateVector
-        Kokkos::parallel_for(Kokkos::RangePolicy<KokkosExecSpace>(0, N),
-                             getProbFunctor<PrecisionT>(arr_data, probability));
-
         // Convert probability distribution to cumulative distribution
+        auto probability = probs_core();
         Kokkos::parallel_scan(
             Kokkos::RangePolicy<KokkosExecSpace>(0, N),
             KOKKOS_LAMBDA(const std::size_t k, PrecisionT &update_value,
@@ -714,17 +684,7 @@ class Measurements final
             Sampler<PrecisionT, Kokkos::Random_XorShift64_Pool>(
                 samples, probability, rand_pool, num_qubits, N));
 
-        std::vector<std::size_t> samples_h(num_samples * num_qubits);
-
-        using UnmanagedSize_tHostView =
-            Kokkos::View<std::size_t *, Kokkos::HostSpace,
-                         Kokkos::MemoryTraits<Kokkos::Unmanaged>>;
-
-        Kokkos::deep_copy(
-            UnmanagedSize_tHostView(samples_h.data(), samples_h.size()),
-            samples);
-
-        return samples_h;
+        return view2vector(samples);
     }
 
   private: